requirements.txt
chardet~=4.0.0
requests~=2.28.1
opencv-python~=4.6.0.66
numpy~=1.23.2
django~=4.1
pillow~=9.2.0
matplotlib~=3.2.2